modify

+++ GAMMA/src/main/java/org/game/user/controller/ConsumerBasketController.java
@@ -47,6 +47,10 @@ public class ConsumerBasketController {
 	public String additionConsumerBasket(long gnum, HttpSession session, Model model) {
 
 		String cid = (String)session.getAttribute("session_cid");
+
+		if(cid == null) {
+			return "redirect:/user/userLogin";
+		}
 		
 		service.additionConsumerBasket(cid, gnum);
 		
@@ -57,6 +61,10 @@ public class ConsumerBasketController {
 	public String removeConsumerBasket(long gnum, HttpSession session, Model model) {
 		
 		String cid = (String)session.getAttribute("session_cid");
+
+		if(cid == null) {
+			return "redirect:/user/userLogin";
+		}
 		
 		service.removeConsumerBasket(cid, gnum);
